+++ slime/mtl/Alg.h
@@ -34,7 +34,7 @@ namespace SLIME {
 
     template<class V, class T>
     static inline void remove(V &ts, const T &t) {
-        int j = 0;
+        long j = 0;
         for (; j < ts.size() && ts[j] != t; j++);
         assert(j < ts.size());
         for (; j < ts.size() - 1; j++) ts[j] = ts[j + 1];
@@ -44,7 +44,7 @@ namespace SLIME {
 
     template<class V, class T>
     static inline bool find(V &ts, const T &t) {
-        int j = 0;
+        long j = 0;
         for (; j < ts.size() && ts[j] != t; j++);
         return j < ts.size();
     }
@@ -65,7 +65,7 @@ namespace SLIME {
     static inline void copy(const vec<T> &from, vec<T> &to, bool append = false) {
         if (!append)
             to.clear();
-        for (int i = 0; i < from.size(); i++) {
+        for (long i = 0; i < from.size(); i++) {
             to.push();
             copy(from[i], to.last());
         }
